Move all tests into the same folder
