+++ src/__tests__/auth.test.js
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import app from '../index';
+import app from '../routes/index';
 
 describe('Sign Up route', () => {
   it('should create a new user', async () => {
