+++ src/__tests__/follower.test.js
@@ -1,4 +1,4 @@
-import Follower from '../../models/follower.model';
+import Follower from '../models/follower.model';
 
 describe('Follower model', () => {
   test('User id and follower id should be present', async () => {
