+++ src/__tests__/tweet.test.js
@@ -1,11 +1,11 @@
-import TweetReply from '../../models/tweetreply.model';
+import Tweet from '../models/tweet.model';
 
 describe('Tweet model', () => {
   test('Body must be required', async () => {
     expect.assertions(1);
 
     try {
-      await TweetReply.create({ });
+      await Tweet.create({ });
     } catch (e) {
       expect(e).toBeTruthy();
     }
@@ -14,7 +14,7 @@ describe('Tweet model', () => {
     expect.assertions(0);
 
     try {
-      await TweetReply.create({
+      await Tweet.create({
         body: "The boy is good"
       });
     } catch (e) {
